backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
google-generativeai>=0.3.0
pillow>=10.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
requests>=2.31.0